rename NodeProcessor -> EntityProcessor
