+++ include/Solver.h
@@ -5,7 +5,7 @@
 
 #include "IoMesh.h"
 #include "MeshData.h"
-#include "NodeProcessor.h"
+#include "EntityProcessor.h"
 
 namespace aperi {
