+++ src/BoundaryCondition.cpp
@@ -5,7 +5,7 @@
 #include "IoInputFile.h"
 #include "MathUtils.h"
 #include "MeshData.h"
-#include "NodeProcessor.h"
+#include "EntityProcessor.h"
 
 namespace aperi {
