+++ src/InitialConditionUtil.cpp
@@ -6,7 +6,7 @@
 #include "LogUtils.h"
 #include "MathUtils.h"
 #include "MeshData.h"
-#include "NodeProcessor.h"
+#include "EntityProcessor.h"
 
 namespace aperi {
