+++ test/BoundaryConditionTest.cpp
@@ -9,7 +9,7 @@
 #include "InitialConditionUtil.h"
 #include "MathUtils.h"
 #include "MeshData.h"
-#include "NodeProcessor.h"
+#include "EntityProcessor.h"
 #include "yaml-cpp/yaml.h"
 
 // Functor to update the nodal displacements
