+++ test/UnitTestUtils.cpp
@@ -18,7 +18,7 @@
 #include "IoMesh.h"
 #include "MathUtils.h"
 #include "MeshData.h"
-#include "NodeProcessor.h"
+#include "EntityProcessor.h"
 
 YAML::Node CreateTestYaml() {
     // This is the yaml file that should be generated by the following input file:
